Move Other class out of Program.cs, rename to OtherTests
